fix: ensure display_order column exists before query and tighten migration error handling
- Extract display_order/is_pinned runtime migration into its own module
  (conversation-display-order-migration.ts) to avoid circular dependencies
  between conversation-store.ts and conversation-queries.ts
- Call ensureDisplayOrderMigration() at the start of listConversations()
  so the column always exists before the ORDER BY references it
- Only ignore "duplicate column name" SQLite errors in the migration;
  log and rethrow any other errors so the ensured flag isn't set incorrectly
