+++ assistant/src/memory/conversation-queries.ts
@@ -4,12 +4,14 @@ import { getLogger } from '../util/logger.js';
 import { getDb, rawAll } from './db.js';
 import { parseConversation, parseMessage } from './conversation-crud.js';
 import type { ConversationRow, MessageRow } from './conversation-crud.js';
+import { ensureDisplayOrderMigration } from './conversation-display-order-migration.js';
 import { conversations, messages } from './schema.js';
 import { buildFtsMatchQuery } from './search/lexical.js';
 
 const log = getLogger('conversation-store');
 
 export function listConversations(limit?: number, includeBackground = false, offset = 0): ConversationRow[] {
+  ensureDisplayOrderMigration();
   const db = getDb();
   const where = includeBackground ? undefined : sql`${conversations.threadType} != 'background'`;
   const query = db
